Support keys and configs in SplitClientForTest
SplitClientForTest can be useful for unit-testing code that depends on
SplitClient, especially in dynamic scenarios where a static config per
test class won't work. As of now, SplitClientForTest doesn't support
customization of split keys or configs - it always returns the same
treatment for all keys of a feature flag and always returns null as a
split config.

This change extends SplitClientForTest, so it's possible to set
treatments for each feature flag and key, and also possible to set split
config in addition to the treatment name.